requirements.txt
websockets
flask
gpiozero
mpu6050-raspberrypi